pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "whale-scanner"
version = "0.1.0"
description = "Unusual options flow scanner — Agent A rule engine for Unusual Whales API"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "pydantic>=2.5",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "structlog>=24.0",
    "aiosqlite>=0.20",
    "asyncio-throttle>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "respx>=0.21",
    "ruff>=0.4",
    "mypy>=1.10",
]
notify = [
    "slack-sdk>=3.27",
]

[project.scripts]
whale-scanner = "scanner.main:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.mypy]
python_version = "3.11"
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
